requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
Django==2.0.13
django-cors-headers==3.1.0
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.7
get==2019.4.13
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
mccabe==0.6.1
nltk==3.4.5
numpy==1.17.2
Pillow==6.0.0
pipenv==2018.11.26
post==2019.4.13
public==2019.4.13
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
style==1.1.0
update==0.0.1
urllib3==1.25.3
virtualenv==16.7.4
virtualenv-clone==0.5.3
